removed old ListView code and supporting classes in favor of newer RecyclerView impl
- also updated GetPlaylistAsyncTask so that it has a pubic constructor
through which the YouTube data api class can be set rather than passing
it in as a parameter in execute()
- updated YouTubeRecyclerViewFragment with a
fragment.setYouTubeDataApi(YouTube) method through which the
newInstance() method can set the YouTube data api instance on the
fragment
