+++ app/src/main/java/com/akoscz/youtube/PlaylistCardAdapter.java
@@ -10,7 +10,7 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.akoscz.youtube.model_v3.Playlist;
+import com.akoscz.youtube.model.Playlist;
 import com.google.api.services.youtube.model.Video;
 import com.google.api.services.youtube.model.VideoContentDetails;
 import com.google.api.services.youtube.model.VideoSnippet;
