--- app/src/main/res/layout/youtube_list_view_fragment.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:weightSum="1"
-    android:background="#ddd"
-    android:orientation="vertical">
-
-    <ListView
-        android:id="@+id/youtube_listview"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:divider="@null"
-        android:layout_weight="1" />
-
-</LinearLayout>
