--- app/src/main/res/layout/youtube_video_list_item.xml
@@ -1,174 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/video_card"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content">
-
-    <RelativeLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        android:layout_marginLeft="12dp"
-        android:layout_marginRight="12dp"
-        android:background="@drawable/rounded_corner">
-
-        <ImageView
-            android:id="@+id/video_thumbnail"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:scaleType="fitXY"
-            android:adjustViewBounds="true"
-            android:layout_alignParentTop="true"
-            tools:src="@drawable/ic_thumb_up"/>
-
-        <TextView
-            android:id="@+id/video_title"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:layout_marginLeft="12dp"
-            android:layout_marginRight="12dp"
-            android:textColor="@android:color/primary_text_light"
-            android:layout_gravity="center_vertical"
-            android:textSize="18sp"
-            android:fontFamily="sans-serif-light"
-            android:layout_below="@id/video_thumbnail"
-            tools:text="Video Title text that is long enough to make the line wrap around to the next line."
-            />
-
-        <RelativeLayout
-            android:id="@+id/video_stats_line"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="12dp"
-            android:layout_marginRight="12dp"
-            android:layout_below="@id/video_title">
-
-            <TextView
-                android:id="@+id/video_view_count"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentRight="true"
-                android:fontFamily="sans-serif-light"
-                android:textSize="20sp"
-                android:textColor="@android:color/primary_text_light"
-                tools:text="119,313"/>
-
-            <ImageView
-                android:id="@+id/video_share"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/video_view_count"
-                android:src="@drawable/ic_share" />
-
-            <TextView
-                android:id="@+id/video_share_text"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/video_view_count"
-                android:layout_toRightOf="@id/video_share"
-                android:layout_alignBottom="@id/video_share"
-                android:gravity="center_vertical"
-                android:layout_marginLeft="8dp"
-                android:fontFamily="sans-serif-light"
-                android:textSize="16sp"
-                android:textColor="@android:color/primary_text_light"
-                android:text="Share"/>
-
-            <ImageView
-                android:id="@+id/video_thumb_up"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/video_view_count"
-                android:layout_toLeftOf="@+id/video_like_count"
-                android:layout_alignBottom="@id/video_share"
-                android:gravity="center_vertical"
-                android:src="@drawable/ic_thumb_up" />
-
-            <TextView
-                android:id="@+id/video_like_count"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/video_view_count"
-                android:layout_toLeftOf="@+id/video_thumb_down"
-                android:layout_alignBottom="@id/video_share"
-                android:gravity="center_vertical"
-                android:layout_marginLeft="12dp"
-                android:layout_marginRight="16dp"
-                android:fontFamily="sans-serif-light"
-                android:textSize="14sp"
-                android:textColor="@android:color/primary_text_light"
-                tools:text="99"/>
-
-            <ImageView
-                android:id="@+id/video_thumb_down"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/video_view_count"
-                android:layout_toLeftOf="@+id/video_dislike_count"
-                android:layout_alignBottom="@id/video_share"
-                android:gravity="center_vertical"
-                android:src="@drawable/ic_thumb_down" />
-
-            <TextView
-                android:id="@+id/video_dislike_count"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_below="@id/video_view_count"
-                android:layout_alignParentRight="true"
-                android:layout_marginLeft="12dp"
-                android:layout_alignBottom="@id/video_share"
-                android:gravity="center_vertical"
-                android:fontFamily="sans-serif-light"
-                android:textSize="14sp"
-                android:textColor="@android:color/primary_text_light"
-                tools:text="99"/>
-
-        </RelativeLayout>
-
-        <LinearLayout
-            android:id="@+id/video_duration"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:layout_marginLeft="12dp"
-            android:layout_marginRight="12dp"
-            android:layout_marginBottom="12dp"
-            android:layout_alignParentRight="true"
-            android:background="@drawable/semi_transparent_box_with_border"
-            android:layout_above="@id/video_title">
-
-            <TextView
-                android:id="@+id/video_dutation_text"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:textColor="@android:color/primary_text_dark"
-                android:fontFamily="sans-serif-light"
-                android:text="--:--"
-                android:paddingLeft="8dp"
-                android:paddingRight="8dp"
-                android:paddingTop="4dp"
-                android:paddingBottom="4dp"
-                android:gravity="center"
-                android:layout_gravity="center"/>
-        </LinearLayout>
-
-        <TextView
-            android:id="@+id/video_description"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="2dp"
-            android:layout_marginBottom="8dp"
-            android:layout_marginLeft="12dp"
-            android:layout_marginRight="12dp"
-            android:textColor="@android:color/secondary_text_light"
-            android:layout_below="@id/video_stats_line"
-            android:fontFamily="sans-serif-light"
-            tools:text="This is a sample description of the video item that is long enough to wrap around to the next line." />
-
-    </RelativeLayout>
-
-</FrameLayout>
